Add Lombok dependency and update controllers and models

+++ src/main/java/com/monlau/springboot/repository/ProductRepository.java
@@ -1,7 +1,10 @@
 package com.monlau.springboot.repository;
 import com.monlau.springboot.model.Product;
+import com.monlau.springboot.model.Order;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+
+public interface ProductRepository extends JpaRepository<Product, Long> {
     
 }
